Corrected some issues found by 'findbugs'. Added 'random steal' type test application.
git-svn-id: http://gforge.cs.vu.nl/svn/ibis/smartsockets/trunk@6209 aaf88347-d911-0410-b711-e54d386773bb

+++ src/test/direct/simple/MultiplexTest2.java
@@ -111,7 +111,7 @@ public class MultiplexTest2 {
 
             time = System.currentTimeMillis() - time;
             
-            long size = TOTAL_SIZE*senders;
+            long size = ((long) TOTAL_SIZE)*senders;
             double sec = time/1000.0;
             double TP = (size / (1024.0*1024.0))/sec;
